% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HairEyeColor_jp}
\alias{HairEyeColor_jp}
\title{日本語化した HairEyeColor データ}
\format{
An object of class \code{table} of dimension 4 x 4 x 2.
}
\usage{
HairEyeColor_jp
}
\description{
日本語化した HairEyeColor データ
}
\keyword{datasets}
