% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_jp.R
\name{mosaic_jp1}
\alias{mosaic_jp1}
\title{colset is the col name of RColorbrewer ex: "Pastel1", "Blues"
library(RColorBrewer)
display.brewer.all(n=10, exact.n=FALSE) # you can see the color palette}
\usage{
mosaic_jp1(.tbl, mtitle = "mosaic_jp", colset = "Set3")
}
\description{
colset is the col name of RColorbrewer ex: "Pastel1", "Blues"
library(RColorBrewer)
display.brewer.all(n=10, exact.n=FALSE) # you can see the color palette
}
