% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_jp.R
\name{mosaic_jp1cell}
\alias{mosaic_jp1cell}
\title{このあと、
labeling_cells(text = round(100*prop.table(.tbl,1),1))(.tbl)　のように表示するデータを指示する。}
\usage{
mosaic_jp1cell(.tbl, mtitle = "mosaic_jp", colset = "Set3")
}
\description{
このあと、
labeling_cells(text = round(100*prop.table(.tbl,1),1))(.tbl)　のように表示するデータを指示する。
}
